% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.gof_mlergm.R
\name{is.gof_mlergm}
\alias{is.gof_mlergm}
\title{Check if object is of class \code{gof_mlergm}}
\usage{
is.gof_mlergm(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
\code{TRUE} if the provided object \code{x} is of class \code{gof_mlergm}, \code{FALSE} otherwise.
}
\description{
Function checks if a provided object is of class \code{gof_mlergm} (see \code{\link{gof.mlergm}} for details).
}
\seealso{
\code{\link{mlergm}}, \code{\link{gof.mlergm}}
}
