% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.mlergm.R
\name{is.mlergm}
\alias{is.mlergm}
\title{Check if the object is of class \code{mlergm}}
\usage{
is.mlergm(x)
}
\arguments{
\item{x}{An objected to be checked.}
}
\value{
\code{TRUE} if the provided object \code{x} is of class \code{mlergm}, \code{FALSE} otherwise.
}
\description{
Function checks if a provided object is of class \code{mlergm} (see \code{\link[mlergm]{mlergm}} for details).
}
\seealso{
\code{\link[mlergm]{mlergm}}
}
