% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.mlnet.R
\name{is.mlnet}
\alias{is.mlnet}
\title{Check if object is of class \code{mlnet}}
\usage{
is.mlnet(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
\code{TRUE} if the provided object \code{x} is of class \code{mlnet}, \code{FALSE} otherwise.
}
\description{
Function checks if a provided object is of class \code{mlnet} (see \code{\link[mlergm]{mlnet}} for details).
}
\seealso{
\code{\link[mlergm]{mlnet}}
}
