% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gof_mlergm.R
\name{print.gof_mlergm}
\alias{print.gof_mlergm}
\title{Print summary of a \code{gof_mlergm} object.}
\usage{
\method{print}{gof_mlergm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gof_mlergm}, probably produced by \code{\link{gof.mlergm}}.}

\item{\dots}{Additional arguments to be passed if necessary.}
}
\description{
Prints a formatted summary output for \code{gof_mlergm} object which was produced by \code{\link{gof.mlergm}}.
}
\seealso{
\code{\link{gof.mlergm}}
}
